requirements.txt
pandas
pydantic
langchain-community
langchain-core
python-dotenv 